+++ saffier/core/files/__init__.py
@@ -1,13 +1,15 @@
 from . import locks
-from .base import ContentFile, File, FileUpload
+from .base import ContentFile, FieldFile, File, FileUpload, ImageFieldFile
 from .move import file_move_safe
 from .storage import FileSystemStorage, Storage, StorageHandler, storages
 
 __all__ = [
     "ContentFile",
+    "FieldFile",
     "File",
     "FileSystemStorage",
     "FileUpload",
+    "ImageFieldFile",
     "Storage",
     "StorageHandler",
     "file_move_safe",